Revert Location and geofencing screens
